chore: Move yggdrasil systemd service management to own package
* Card ID: CCT-1584

To improve code separation, move the yggdrasil methods to their own
package.
